Reader achievements: move shared components to client/reader/components/achievements
