+++ client/reader/components/achievements/years-of-service-badge/style.scss
